feat: Elevator state machine and common Elevator states
